requirements.txt
torch>=1.10.0
torchvision>=0.11.0
timm
transformers
numpy
Pillow
tqdm
pyyaml
matplotlib
pandas
